Reworked the file structure
